# server-go/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.ProductInput:
    properties:
      categoryId:
        type: integer
      description:
        type: string
      imageUrl:
        type: string
      images:
        type: string
      isActive:
        type: boolean
      isFeatured:
        type: boolean
      name:
        type: string
      price:
        type: number
      slug:
        type: string
      stock:
        type: integer
    type: object
...
  models.Category:
    properties:
      children:
        items:
          $ref: '#/definitions/models.Category'
        type: array
      createdAt:
        type: string
      id:
        type: integer
      name:
        type: string
      parent:
        $ref: '#/definitions/models.Category'
      parentId:
        type: integer
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      slug:
        type: string
      updatedAt:
        type: string
    type: object
...
  models.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  models.Product:
    properties:
      category:
        $ref: '#/definitions/models.Category'
      categoryId:
        type: integer
      createdAt:
        type: string
      description:
        type: string
      id:
        type: integer
      imageUrl:
        type: string
      images:
        type: string
      isActive:
        type: boolean
      isFeatured:
        type: boolean
      name:
        type: string
      price:
        type: number
      slug:
        type: string
      stock:
        type: integer
      updatedAt:
        type: string
    type: object
  models.RegisterInput:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
...
  models.UpdateProfileInput:
    properties:
      name:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: support@bycigar.com
    name: API Support
  description: BYCIGAR 电商后台 API 接口文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: BYCIGAR API
  version: "1.0"
paths:
...
  /admin/products:
    get:
      consumes:
      - application/json
      description: 获取商品列表，支持分页和筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 搜索关键词
        in: query
        name: search
        type: string
      - description: 分类ID
        in: query
        name: categoryId
        type: integer
      - description: 是否推荐
        in: query
        name: featured
        type: string
      - description: 是否上架
        in: query
        name: active
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取商品列表(管理端)
      tags:
      - admin-products
    post:
      consumes:
      - application/json
      description: 创建新的商品
      parameters:
      - description: 商品信息
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.ProductInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 创建商品
      tags:
      - admin-products
  /admin/products/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定商品
      parameters:
      - description: 商品ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 删除商品
      tags:
      - admin-products
    put:
      consumes:
      - application/json
      description: 更新指定商品信息
      parameters:
      - description: 商品ID
        in: path
        name: id
        required: true
        type: integer
      - description: 商品信息
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handlers.ProductInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新商品
      tags:
      - admin-products
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: 用户登录获取token
      parameters:
      - description: 登录信息
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      summary: 用户登录
      tags:
      - auth
  /auth/me:
    get:
      consumes:
      - application/json
      description: 获取当前登录用户的个人信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 获取用户信息
      tags:
      - auth
  /auth/profile:
    put:
      consumes:
      - application/json
      description: 更新当前用户的个人信息
      parameters:
      - description: 用户信息
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.UpdateProfileInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: 更新用户信息
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: 注册新用户账号
      parameters:
      - description: 注册信息
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/models.RegisterInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: 用户注册
      tags:
      - auth
...
  /products:
    get:
      consumes:
      - application/json
      description: 获取产品列表，支持分页、搜索、分类筛选和排序
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 12
        description: 每页数量
        in: query
        name: limit
        type: integer
      - description: 搜索关键词
        in: query
        name: search
        type: string
      - description: 分类 slug
        in: query
        name: category
        type: string
      - description: 分类 ID
        in: query
        name: categoryId
        type: integer
      - description: 是否精选
        in: query
        name: featured
        type: string
      - default: createdAt
        description: 排序字段
        in: query
        name: sortBy
        type: string
      - default: desc
        description: 排序方向 (asc/desc)
        in: query
        name: sortOrder
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 获取产品列表
      tags:
      - products
  /products/{id}:
    get:
      consumes:
      - application/json
      description: 根据ID获取产品详情
      parameters:
      - description: 产品ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Product'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: 获取单个产品
      tags:
      - products
schemes:
- http
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"